app/services/earthquake/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.3
flask-swagger-ui==4.11.1
python-dotenv==1.0.1 